Fixed log time, only logging after discard time, fixed bug.
Implemented beta function, still need to test more.
Print beta monte carlo is commented.

Implemented nasch with beta.

+++ src/tuff/engine/FDPProvider.java
@@ -15,7 +15,7 @@ public abstract class FDPProvider {
         return false;
     }
 
-    public float provide(float a, float b) {
+    public float provide(int a, int b) {
         return 1;
 
     }
